separate log parser from ingester
