+++ contracts/sdk/indexer/indexer/log/gummyroll.ts
@@ -1,17 +1,13 @@
 import * as anchor from '@project-serum/anchor';
 import { PublicKey } from '@solana/web3.js';
-import { Gummyroll, PathNode } from "../../gummyroll"
+import { Gummyroll } from "../../../gummyroll"
+import { PROGRAM_ID as GUMMYROLL_PROGRAM_ID } from '../../../gummyroll';
 import { parseEventFromLog, ParsedLog, ixRegEx } from './utils';
+import { ParserState } from '../utils';
+import { ChangeLogEvent } from '../ingester';
 
 export type GummyrollIx =
     'InitEmptyGummyroll' | 'InitEmptyGummyrollWithRoot' | 'Replace' | 'Append' | 'InsertOrAppend' | 'VerifyLeaf' | 'TransferAuthority';
-
-export type ChangeLogEvent = {
-    id: PublicKey,
-    path: PathNode[],
-    seq: number,
-    index: number,
-};
 
 function parseIxName(logLine: string): GummyrollIx | null {
     return logLine.match(ixRegEx)[1] as GummyrollIx
@@ -33,4 +29,20 @@ export function parseEventGummyroll(parsedLog: ParsedLog, gummyroll: anchor.Prog
 
 function parseChangelogEvent(logs: string[], gummyroll: anchor.Program<Gummyroll>): ChangeLogEvent | null {
     return parseEventFromLog(logs[logs.length - 2], gummyroll.idl).data as ChangeLogEvent;
+}
+
+export function findGummyrollEvent(
+    logs: (string | ParsedLog)[],
+    parser: ParserState
+): ChangeLogEvent | null {
+    let changeLog: ChangeLogEvent | null;
+    for (const log of logs) {
+        if (typeof log !== "string" && log.programId.equals(GUMMYROLL_PROGRAM_ID)) {
+            changeLog = parseEventGummyroll(log, parser.Gummyroll);
+        }
+    }
+    if (!changeLog) {
+        console.log("Failed to find gummyroll changelog");
+    }
+    return changeLog;
 }
